+++ tests/test_logging.py
@@ -55,16 +55,15 @@ class TestProfileLogging(TestCase):
         log_capture_string = FIFOIO(256)
         ch = logging.StreamHandler(log_capture_string)
         ch.setFormatter(logging.Formatter('%(asctime)s.%(msecs)03d %(levelname)s %(name)s - %(message)s'))
-        getLogger('profile').setLevel('DEBUG')
-        getLogger('profile').addHandler(ch)
+        getLogger('ocrd.process.profile').setLevel('DEBUG')
+        getLogger('ocrd.process.profile').addHandler(ch)
 
         run_processor(DummyProcessor, resolver=Resolver(), mets_url=assets.url_of('SBB0000F29300010000/data/mets.xml'))
 
         log_contents = log_capture_string.getvalue()
         log_capture_string.close()
-        # Check whether profile information has been logged. Dummy should
-        # finish in under 0.1s
-        self.assertTrue(match(r'.*profile.ocrd.processor.ocrd-test - 0.0\d+s.*', log_contents))
+        # Check whether profile information has been logged. Dummy should finish in under 0.1s
+        self.assertTrue(match(r'.*Executing processor "ocrd-test" took 0.\d+s.*', log_contents))
 
 if __name__ == '__main__':
     main()